feat: extract chat component
